requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
sentence-transformers>=2.0.0
firebase-admin>=6.0.0
requests>=2.28.0
python-decouple>=3.6